% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsVKS-lite.r
\name{avg.model}
\alias{avg.model}
\title{avg.model}
\usage{
avg.model(x)
}
\arguments{
\item{x}{= vector of results}
}
\description{
constant-only (zero-variable) regression model
}
\seealso{
Other avg: 
\code{\link{avg.winsorized}()},
\code{\link{avg.wtd}()}
}
\concept{avg}
\keyword{avg.model}
