% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsVKS-lite.r
\name{char.to.int}
\alias{char.to.int}
\title{char.to.int}
\usage{
char.to.int(x)
}
\arguments{
\item{x}{= a character vector}
}
\description{
ascii values
}
\seealso{
Other char: 
\code{\link{char.ex.int}()},
\code{\link{char.lag}()},
\code{\link{char.seq}()},
\code{\link{char.to.num}()}
}
\concept{char}
\keyword{char.to.int}
