% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsVKS-lite.r
\name{combinations}
\alias{combinations}
\title{combinations}
\usage{
combinations(x, y)
}
\arguments{
\item{x}{= a vector}

\item{y}{= an integer between 1 and <length(x)>}
}
\description{
returns all possible combinations of <y> values of <x>
}
\seealso{
Other combinations: 
\code{\link{combinations.ex.int}()},
\code{\link{combinations.next}()},
\code{\link{combinations.to.int}()}
}
\concept{combinations}
\keyword{combinations}
