% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsVKS-lite.r
\name{cpt.RgnSec}
\alias{cpt.RgnSec}
\title{cpt.RgnSec}
\usage{
cpt.RgnSec(x, y)
}
\arguments{
\item{x}{= a vector of Sectors}

\item{y}{= a vector of country codes}
}
\description{
makes Region-Sector groupings
}
\seealso{
Other cpt: 
\code{\link{cpt.RgnSecJP}()}
}
\concept{cpt}
\keyword{cpt.RgnSec}
