% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsVKS-lite.r
\name{Ctry.to.CtryGrp}
\alias{Ctry.to.CtryGrp}
\title{Ctry.to.CtryGrp}
\usage{
Ctry.to.CtryGrp(x)
}
\arguments{
\item{x}{= a vector of country codes}
}
\description{
makes Country groups
}
\seealso{
Other Ctry: 
\code{\link{Ctry.info}()},
\code{\link{Ctry.msci.index.changes}()},
\code{\link{Ctry.msci.members.rng}()},
\code{\link{Ctry.msci.members}()},
\code{\link{Ctry.msci}()}
}
\concept{Ctry}
\keyword{Ctry.to.CtryGrp}
