% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsVKS-lite.r
\name{email.exists}
\alias{email.exists}
\title{email.exists}
\usage{
email.exists(x, y)
}
\arguments{
\item{x}{= report name}

\item{y}{= date for which you want to send the report}
}
\description{
T/F depending on whether email already went out
}
\seealso{
Other email: 
\code{\link{email.kill}()},
\code{\link{email.list}()},
\code{\link{email.record}()}
}
\concept{email}
\keyword{email.exists}
