% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsVKS-lite.r
\name{gram.schmidt}
\alias{gram.schmidt}
\title{gram.schmidt}
\usage{
gram.schmidt(x, y)
}
\arguments{
\item{x}{= a numeric vector/matrix/data frame}

\item{y}{= a numeric isomekic vector}
}
\description{
Gram-Schmidt orthogonalization of <x> to <y>
}
\keyword{gram.schmidt}
