% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsVKS-lite.r
\name{html.list}
\alias{html.list}
\title{html.list}
\usage{
html.list(x)
}
\arguments{
\item{x}{= a string vector}
}
\description{
<x> expressed as an html list
}
\seealso{
Other html: 
\code{\link{html.and}()},
\code{\link{html.ex.utf8}()},
\code{\link{html.flow.breakdown}()},
\code{\link{html.flow.english}()},
\code{\link{html.flow.underlying}()},
\code{\link{html.image}()},
\code{\link{html.positioning}()},
\code{\link{html.problem.underlying}()},
\code{\link{html.problem}()},
\code{\link{html.signature}()},
\code{\link{html.tbl}()},
\code{\link{html.tenure}()}
}
\concept{html}
\keyword{html.list}
