% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsVKS-lite.r
\name{latin.to.arabic}
\alias{latin.to.arabic}
\title{latin.to.arabic}
\usage{
latin.to.arabic(x)
}
\arguments{
\item{x}{= a character vector of latin numerals}
}
\description{
returns <x> expressed as an integer
}
\seealso{
Other latin: 
\code{\link{latin.ex.arabic}()},
\code{\link{latin.to.arabic.underlying}()}
}
\concept{latin}
\keyword{latin.to.arabic}
