% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsVKS-lite.r
\name{principal.components}
\alias{principal.components}
\title{principal.components}
\usage{
principal.components(x, y = 2)
}
\arguments{
\item{x}{= a matrix/df}

\item{y}{= number of principal components desired}
}
\description{
first <y> principal components
}
\seealso{
Other principal: 
\code{\link{principal.components.covar}()},
\code{\link{principal.components.underlying}()}
}
\concept{principal}
\keyword{principal.components}
