% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsVKS-lite.r
\name{publish.daily.last}
\alias{publish.daily.last}
\title{publish.daily.last}
\usage{
publish.daily.last(x)
}
\arguments{
\item{x}{= a YYYYMMDD date}
}
\description{
last daily flow-publication date
}
\seealso{
Other publish: 
\code{\link{publish.monthly.last}()},
\code{\link{publish.weekly.last}()}
}
\concept{publish}
\keyword{publish.daily.last}
