% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsVKS-lite.r
\name{qtr.ex.int}
\alias{qtr.ex.int}
\title{qtr.ex.int}
\usage{
qtr.ex.int(x)
}
\arguments{
\item{x}{= a vector of integers}
}
\description{
returns a vector of <yyyymm> months
}
\seealso{
Other qtr: 
\code{\link{qtr.lag}()},
\code{\link{qtr.seq}()},
\code{\link{qtr.to.int}()}
}
\concept{qtr}
\keyword{qtr.ex.int}
