% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsVKS-lite.r
\name{qtr.to.int}
\alias{qtr.to.int}
\title{qtr.to.int}
\usage{
qtr.to.int(x)
}
\arguments{
\item{x}{= a vector of <qtr>}
}
\description{
returns a vector of integers
}
\seealso{
Other qtr: 
\code{\link{qtr.ex.int}()},
\code{\link{qtr.lag}()},
\code{\link{qtr.seq}()}
}
\concept{qtr}
\keyword{qtr.to.int}
