% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsVKS-lite.r
\name{read.EPFR}
\alias{read.EPFR}
\title{read.EPFR}
\usage{
read.EPFR(x)
}
\arguments{
\item{x}{= a path to a file written by the dev team}
}
\description{
reads in the file
}
\seealso{
Other read: 
\code{\link{read.prcRet}()}
}
\concept{read}
\keyword{read.EPFR}
