% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsVKS-lite.r
\name{sqlts.FloDollar.daily}
\alias{sqlts.FloDollar.daily}
\title{sqlts.FloDollar.daily}
\usage{
sqlts.FloDollar.daily(x)
}
\arguments{
\item{x}{= the security id for which you want data}
}
\description{
SQL query for daily dollar flow
}
\seealso{
Other sqlts: 
\code{\link{sqlts.FloDollar.monthly}()},
\code{\link{sqlts.TopDownAllocs}()}
}
\concept{sqlts}
\keyword{sqlts.FloDollar.daily}
