% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsVKS-lite.r
\name{strat.file}
\alias{strat.file}
\title{strat.file}
\usage{
strat.file(x, y)
}
\arguments{
\item{x}{= vector of strategy names (e.g. "FX" or "FloPctCtry")}

\item{y}{= frequency (e.g. "daily", "weekly" or "monthly")}
}
\description{
the path to the factor file
}
\seealso{
Other strat: 
\code{\link{strat.dir}()},
\code{\link{strat.list}()},
\code{\link{strat.path}()}
}
\concept{strat}
\keyword{strat.file}
